muOS Internals - Added network loopback device - Fixed HDMI sound not being reset on reboot - Fixed RetroArch command line issue - Fixed application launchers to use module variable system - Fixed audio channel in device specific configuration - Fixed file permissions - Fixed input mapping state machine - Fixed random charging state skip issue - Fixed root mount location for RG28XX - Fixed screenshot being triggered due to stuck power key input - Fixed tasks to use module variable system - Moved all scripts to a modular variable system - Moved input combo states to simplified case switch

+++ init/MUOS/task/Clear Favourites.sh
@@ -1,44 +1,29 @@
 #!/bin/sh
 
-# Grab device variables
-. /opt/muos/script/system/parse.sh
-DEVICE=$(tr '[:upper:]' '[:lower:]' < "/opt/muos/config/device.txt")
-DEVICE_CONFIG="/opt/muos/device/$DEVICE/config.ini"
+. /opt/muos/script/var/func.sh
 
-ROM_MOUNT=$(parse_ini "$DEVICE_CONFIG" "storage.rom" "mount")
+. /opt/muos/script/var/device/storage.sh
 
-# Suspend the muxtask program
 pkill -STOP muxtask
-
-# Fire up the logger!
 /opt/muos/extra/muxlog &
-sleep 1
-
-echo "Waiting..." > /tmp/muxlog_info
 sleep 1
 
 TMP_FILE=/tmp/muxlog_global
 rm -rf "$TMP_FILE"
 
-# Grab current date
-DATE=$(date +%Y-%m-%d)
+MUOS_FAV_DIR="$DC_STO_ROM_MOUNT/MUOS/info/favourite"
 
-# muOS Favourites Directory
-MUOS_FAV_DIR="$ROM_MOUNT/MUOS/info/favourite"
+echo "Deleting Favourite Files" >/tmp/muxlog_info
+rm -rf "${MUOS_FAV_DIR:?}"/*
 
-echo "Deleting Favourite Files" > /tmp/muxlog_info
-rm -rf "$MUOS_FAV_DIR"/*
-
-# Sync filesystem just-in-case :)
-echo "Sync Filesystem" > /tmp/muxlog_info
+echo "Sync Filesystem" >/tmp/muxlog_info
 sync
 
-echo "All Done!" > /tmp/muxlog_info
+echo "All Done!" >/tmp/muxlog_info
 sleep 1
 
 killall -q muxlog
 rm -rf "$MUX_TEMP" /tmp/muxlog_*
 
-# Resume the muxtask program
 pkill -CONT muxtask
 killall -q "Clear Favourites.sh"
