+++ init/MUOS/task/Clear System Configuration.sh
@@ -1,47 +1,32 @@
 #!/bin/sh
 
-# Grab device variables
-. /opt/muos/script/system/parse.sh
-DEVICE=$(tr '[:upper:]' '[:lower:]' < "/opt/muos/config/device.txt")
-DEVICE_CONFIG="/opt/muos/device/$DEVICE/config.ini"
+. /opt/muos/script/var/func.sh
 
-ROM_MOUNT=$(parse_ini "$DEVICE_CONFIG" "storage.rom" "mount")
+. /opt/muos/script/var/device/storage.sh
 
-# Suspend the muxtask program
 pkill -STOP muxtask
-
-# Fire up the logger!
 /opt/muos/extra/muxlog &
-sleep 1
-
-echo "Waiting..." > /tmp/muxlog_info
 sleep 1
 
 TMP_FILE=/tmp/muxlog_global
 rm -rf "$TMP_FILE"
 
-# Grab current date
-DATE=$(date +%Y-%m-%d)
+MUOS_CONFIG_DIR="$DC_STO_ROM_MOUNT/MUOS/info/config"
 
-# muOS Favourites Directory
-MUOS_CONFIG_DIR="$ROM_MOUNT/MUOS/info/config"
+echo "Removing all configurations" >/tmp/muxlog_info
+rm -rf "${MUOS_CONFIG_DIR:?}"/*
 
-echo "Removing all configurations" > /tmp/muxlog_info
-rm -rf "$MUOS_CONFIG_DIR"/*
+echo "Adding default RetroArch shader" >/tmp/muxlog_info
+echo '#reference "../../retroarch/shaders/shimmerless/sharp-shimmerless.glslp"' >"$MUOS_CONFIG_DIR/global.glslp"
 
-echo "Adding default RetroArch shader" > /tmp/muxlog_info
-echo '#reference "../../retroarch/shaders/shimmerless/sharp-shimmerless.glslp"' > "$MUOS_CONFIG_DIR/global.glslp"
-
-# Sync filesystem just-in-case :)
-echo "Sync Filesystem" > /tmp/muxlog_info
+echo "Sync Filesystem" >/tmp/muxlog_info
 sync
 
-echo "All Done!" > /tmp/muxlog_info
+echo "All Done!" >/tmp/muxlog_info
 sleep 1
 
 killall -q muxlog
 rm -rf "$MUX_TEMP" /tmp/muxlog_*
 
-# Resume the muxtask program
 pkill -CONT muxtask
 killall -q "Clear System Configuration.sh"
